requirements.txt
streamlit
langchain
langchain-community
langchain-chroma
langchain-huggingface
chromadb
sentence-transformers
tqdm
langchain-google-genai
unstructured[pdf]
python-dotenv